python refinements
