+++ python/coinbase-agentkit/scripts/generate_action_provider/main.py
@@ -8,20 +8,25 @@ then generates all necessary files with appropriate boilerplate code.
 @module scripts/create-action-provider
 """
 
+from pathlib import Path
+
 from .args import parse_cli_args
 from .config import prepare_provider_config
-from .utils import create_provider_files, display_banner, display_success_message
+from .utils import (
+    add_provider_files,
+    display_banner,
+    display_success_message,
+    update_action_providers_init,
+    update_agentkit_init,
+)
 
 
 def create_action_provider() -> None:
     """Create a new action provider using CLI args or interactive prompts."""
     display_banner()
 
-    # Parse CLI args
-    name, protocol_family, networks, wallet_provider, interactive = parse_cli_args()
-
     try:
-        # Prepare config with CLI args and prompt fallbacks
+        name, protocol_family, networks, wallet_provider, interactive = parse_cli_args()
         config = prepare_provider_config(
             name=name,
             protocol_family=protocol_family,
@@ -30,8 +35,12 @@ def create_action_provider() -> None:
             interactive=interactive,
         )
 
-        # Generate files
-        create_provider_files(config)
+        provider_dir = Path("coinbase_agentkit/action_providers") / config.name
+        test_dir = Path("tests/action_providers") / config.name
+        add_provider_files(config, provider_dir, test_dir)
+        update_action_providers_init(config)
+        update_agentkit_init(config)
+
         display_success_message(config.name)
 
     except KeyboardInterrupt:
@@ -44,3 +53,7 @@ def create_action_provider() -> None:
 def main() -> None:
     """Execute the main entry point."""
     create_action_provider()
+
+
+if __name__ == "__main__":
+    main()
