+++ python/coinbase-agentkit/scripts/generate_action_provider/templates/README.md.template
@@ -4,16 +4,26 @@ This directory contains the **{{ name_pascal }}ActionProvider** implementation, 
 
 ## Overview
 
-The {{ name_pascal }}ActionProvider is designed to work with the {{ wallet_provider }} for blockchain interactions. It provides a set of actions that enable [describe the main purpose/functionality].
+The {{ name_pascal }}ActionProvider is designed to {% if wallet_provider %}work with {{ wallet_provider }} for blockchain interactions{% else %}provide blockchain-agnostic operations{% endif %}. It provides a set of actions that enable [describe the main purpose/functionality].
 
 ## Directory Structure
 
 ```
 {{ name }}/
-├── {{ name }}_action_provider.py    # Main provider implementation
-├── {{ name }}_action_provider_test.py # Test suite
-├── schemas.py                       # Action schemas and types
-└── README.md                       # Documentation (this file)
+├── __init__.py                    # Package exports
+├── {{ name }}_action_provider.py  # Main provider implementation
+├── schemas.py                     # Action schemas and types
+└── README.md                      # Documentation (this file)
+```
+
+## Test Structure
+
+```
+tests/action_providers/{{ name }}/
+├── __init__.py                  # Test package initialization
+├── conftest.py                  # Test fixtures and configuration
+├── test_action_provider.py      # Provider test suite
+└── test_example_action.py       # Action-specific tests
 ```
 
 ## Actions
@@ -22,35 +32,36 @@ The {{ name_pascal }}ActionProvider is designed to work with the {{ wallet_provi
 - `example-action`: Template action implementation
   - **Purpose**: Demonstrates the basic structure of an action
   - **Input**:
-    - `field_name` (string): A descriptive name for the field
-    - `amount` (string): The amount to use in the action (e.g. "1.5")
+    - `field_name` (string): A descriptive name for the field (1-100 chars)
+    - `amount` (string): The amount as a decimal string (e.g. "1.5")
     - `optional_field` (string, optional): Optional parameter example
-    - `enum_field` ("option1" | "option2" | "option3"): Constrained choices
   - **Output**: String describing the action result
   - **Example**:
     ```python
-    result = await provider.example_action(wallet_provider, {
+    result = provider.example_action({% if wallet_provider %}wallet_provider, {% endif %}{
         "field_name": "test",
-        "amount": "1.0",
-        "enum_field": "option1"
+        "amount": "1.0"
     })
     ```
 
 ## Implementation Details
 
 ### Network Support
-The {{ name_pascal }} Action Provider supports networks based on the {{ wallet_provider }} requirements:
-{% if wallet_provider == "EvmWalletProvider" %}
-- Supports EVM-compatible networks (protocol_family: "evm")
-{% elif wallet_provider == "WalletProvider" %}
-- Supports all networks (generic wallet provider)
+{% if not protocol_family %}
+This provider supports all blockchain networks.
+{% elif network_ids|length == 0 %}
+This provider supports all {{ protocol_family }} networks.
+{% else %}
+This provider supports {{ protocol_family }} networks: {% for networkId in network_ids %}{{ networkId }}{% if not loop.last %}, {% endif %}{% endfor %}.
 {% endif %}
 
+{% if wallet_provider %}
 ### Wallet Provider Integration
 This provider is specifically designed to work with {{ wallet_provider }}. Key integration points:
 - Network compatibility checks
 - Transaction signing and execution
 - Balance and account management
+{% endif %}
 
 ## Adding New Actions
 
@@ -60,7 +71,8 @@ To add new actions:
    ```python
    class NewActionSchema(BaseModel):
        # Define your action's parameters
-       pass
+       field_name: str = Field(...)
+       amount: str = Field(...)
    ```
 
 2. Implement the action in `{{ name }}_action_provider.py`:
@@ -72,7 +84,7 @@ To add new actions:
    )
    def new_action(
        self,
-       wallet_provider: {{ wallet_provider }},
+       {% if wallet_provider %}wallet_provider: {{ wallet_provider }},{% endif %}
        args: dict[str, Any],
    ) -> str:
        # Implement your action logic
@@ -88,9 +100,11 @@ When implementing new actions, ensure to:
 
 ## Dependencies
 - AgentKit core library
+{% if wallet_provider %}
 - {{ wallet_provider }} for blockchain interactions
 {% if wallet_provider == "EvmWalletProvider" %}
 - EVM-specific dependencies (e.g., web3.py)
+{% endif %}
 {% endif %}
 
 ## Notes
