+++ python/coinbase-agentkit/scripts/generate_action_provider/templates/__init__.py.template
