+++ python/coinbase-agentkit/scripts/generate_action_provider/templates/action_provider_test.py.template
