+++ python/coinbase-agentkit/scripts/generate_action_provider/templates/conftest.py.template
