+++ python/coinbase-agentkit/scripts/generate_action_provider/types.py
@@ -3,11 +3,11 @@
 from dataclasses import dataclass
 from typing import Literal, TypedDict
 
-# Protocol family types
-# None indicates support for all networks
-ProtocolFamily = Literal["evm"] | None
+
+ProtocolFamily = Literal["all", "evm"] | None
 NetworkId = str
 WalletProvider = str
+
 
 @dataclass
 class ProviderConfig:
@@ -17,6 +17,7 @@ class ProviderConfig:
     protocol_family: ProtocolFamily
     network_ids: list[NetworkId]
     wallet_provider: WalletProvider | None = None
+
 
 class PromptResult(TypedDict, total=False):
     """Result from the prompts."""
